- Refactor class
