+++ src/config/supported_effects.js
@@ -3,51 +3,4 @@
  */
 
 module.exports = {
-    supportedType: [
-        'Original',
-        'Lark',
-        'Reyes',
-        'Juno',
-        'Slumber',
-        'Crema',
-        'Ludwig',
-        'Aden',
-        'Perpetua',
-        'Amaro',
-        'Mayfair',
-        'Rise',
-        'Hudson',
-        'Valencia',
-        'XProII',
-        'Sierra',
-        'Willow',
-        'Lo-Fi',
-        'Earlybird',
-        'Brannan',
-        'Inkwell',
-        'Hefe',
-        'Nashville',
-        'Sutro',
-        'Toaster',
-        'Walden',
-        '1977',
-        'Kelvin',
-        'Grayscale',
-        'Sepia',
-        'Luminance',
-        'Brighten',
-        'Darken',
-        'Opacity',
-        'Threshold',
-        'Nega-Posi',
-        'BrightnessContrast',
-        'Hue-Rotate',
-        'Saturate',
-        'horizontalFlip',
-        'verticalFlip',
-        'doubleFlip',
-        'horizontalMirror',
-        'verticalMirror',
-        'XYMirror'
-    ]
 };
