+++ src/effects/rgb_map.js
