+++ src/index.js
@@ -1,37 +1,99 @@
-/**
- * index.js
- */
-
 /** @ignore */
 const canvas = require('./canvas/canvas');
 /** @ignore */
 const adapter = require('./adapter/adapter');
 
 /**
- * Apply filter.
+ * Filter root class.
  *
- * @param {Buffer} imageBuffer - a image buffer to appply filter
- * @param {String} type - filter type
- * @param {Object} optionArgs - options
- * @returns {*}
+ * @type {Filter}
  */
-module.exports.filter = (imageBuffer, type, optionArgs) => {
-    const options = !optionArgs ? optionArgs : {};
-
-    let imageData = imageBuffer;
-    let convertResult = {};
-    if (Buffer.isBuffer(imageBuffer)) {
-        convertResult = canvas.convert(imageBuffer);
-        imageData = convertResult.imageData;
+module.exports = class Filter {
+    constructor() {
+        this.effects = [
+            'Original',
+            'Lark',
+            'Reyes',
+            'Juno',
+            'Slumber',
+            'Crema',
+            'Ludwig',
+            'Aden',
+            'Perpetua',
+            'Amaro',
+            'Mayfair',
+            'Rise',
+            'Hudson',
+            'Valencia',
+            'XProII',
+            'Sierra',
+            'Willow',
+            'Lo-Fi',
+            'Earlybird',
+            'Brannan',
+            'Inkwell',
+            'Hefe',
+            'Nashville',
+            'Sutro',
+            'Toaster',
+            'Walden',
+            '1977',
+            'Kelvin',
+            'Grayscale',
+            'Sepia',
+            'Luminance',
+            'Brighten',
+            'Darken',
+            'Opacity',
+            'Threshold',
+            'Nega-Posi',
+            'BrightnessContrast',
+            'Hue-Rotate',
+            'Saturate',
+            'horizontalFlip',
+            'verticalFlip',
+            'doubleFlip',
+            'horizontalMirror',
+            'verticalMirror',
+            'XYMirror'
+        ];
     }
 
-    const func = adapter.handleType(type);
-    const convertedPixels = func.apply(this, [imageData.data, options]);
-    imageData.data = convertedPixels;
+    /**
+     * Get all supported effect types
+     *
+     * @returns {Array|string[]} - supported effect types
+     */
+    getSupportedEffects() {
+        return this.effects;
+    };
 
-    if (convertResult.context) {
-        convertResult.context.putImageData(imageData, 0, 0);
-        return canvas.getBase64(convertResult.canvas);
-    }
-    return imageData;
+    /**
+     * Apply filter.
+     *
+     * @param {Buffer} imageBuffer - a image buffer to apply filter
+     * @param {String} effect - filter effect
+     * @param {Object} optionArgs - options
+     * @returns {*} - filter-applied image
+     */
+    apply(imageBuffer, effect, optionArgs) {
+        const options = !optionArgs ? optionArgs : {};
+
+        let imageData = imageBuffer;
+        let convertResult = {};
+        if (Buffer.isBuffer(imageBuffer)) {
+            convertResult = canvas.convert(imageBuffer);
+            imageData = convertResult.imageData;
+        }
+
+        const func = adapter.effectFunc(effect);
+        const convertedPixels = func.apply(this, [imageData.data, options]);
+        imageData.data = convertedPixels;
+
+        if (convertResult.context) {
+            convertResult.context.putImageData(imageData, 0, 0);
+            return canvas.getBase64(convertResult.canvas);
+        }
+        return imageData;
+    };
 };
